// src/types/businessSettings.ts
/*
|--------------------------------------------------------------------------
| Business Settings
|--------------------------------------------------------------------------
*/

export interface BusinessSettings {
    address: string | undefined;
    email: string | undefined;
    phone: string | undefined;
    business_description: string | undefined;
    business_name: string;

    /*
    |--------------------------------------------------------------------------
    | Identity
    |--------------------------------------------------------------------------
    */

    id?: string;

    client_id: string;

    /*
    |--------------------------------------------------------------------------
    | Appointment Settings
    |--------------------------------------------------------------------------
    */

    appointment_duration: number;

    buffer_minutes: number;

    /*
    |--------------------------------------------------------------------------
    | Business Hours
    |--------------------------------------------------------------------------
    */

    opening_time: string;

    closing_time: string;

    /*
    |--------------------------------------------------------------------------
    | Working Days
    |--------------------------------------------------------------------------
    */

    working_days: string[];

    /*
    |--------------------------------------------------------------------------
    | Timestamps
    |--------------------------------------------------------------------------
    */

    created_at?: string;

    updated_at?: string;

}